Add something to search for when looking for errors, not just color

+++ src/edu/stanford/nlp/ling/tokensregex/parser/TokenSequenceParserTokenManager.java
@@ -808,7 +808,7 @@ private int jjMoveNfa_0(int startState, int curPos)
       }
       else
       {
-         int hiByte = (int)(curChar >> 8);
+         int hiByte = (curChar >> 8);
          int i1 = hiByte >> 6;
          long l1 = 1L << (hiByte & 077);
          int i2 = (curChar & 0xff) >> 6;
